setting TaskExecutionContext OutputReader even if outputs do not exist (#455)
## Overview
This PR sets the output reader on all fasttask executions regardless of whether they have ouput variables declared or not. This ensures that errors are correctly [detected by the TaskHandler](https://github.com/unionai/flyte/blob/daniel%2Fdx-981-errors-do-not-propagate-if-there-is-no-output/flytepropeller/pkg/controller/nodes/task/handler.go#L675-L677) (through an `errors.pb` file in blobstore) for all task definitions. 

## Test Plan
This was tested locally on a variety of task definitions.

## Rollout Plan (if applicable)
May be rolled out immediately.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/DX-981/errors-do-not-propagate-if-there-is-no-output

## Checklist
* [x] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation