Fix handling of TTL seconds for fasttask (#377)
## Overview
This PR updates handling of TTLSeconds configuration for fasttask execution environments. If the `TerminationCriteria` is unset the environment will use a FlytePropeller configuration value. If it is set, then we ensure it is strictly greater than 0.

## Test Plan
Tested locally and unit tested.

## Rollout Plan (if applicable)
This may be rolled out immediately.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/COR-1533/fix-handling-of-unset-ttlseconds-for-fasttask

## Checklist
* [x] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation