fasttask plugin supports actor task types (#458)
## Overview
This PR enables tasks of type `actor` to be executed using the `fast-task` plugin. This means the UI will show `Actor` for the task type rather than `Fast Task`.

## Test Plan
Tested locally.

## Rollout Plan (if applicable)
Must update flytepropeller configmap to enable mapping of the `actor` task type to the `fast-task` plugin. Example:
```
tasks:
  task-plugins:
    enabled-plugins:
      ...
      - fast-task
    default-for-task-types:
      - actor: fast-task
      ....
```

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/COR-1601/update-task-type-to-actor

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation