Sanitizing environment names to fix k8s pods (#456)
## Overview
This PR introduces a executionEnvID sanitization process which ensures k8s pods created by the `InMemoryBuilder` have valid names. This is one proposal to fix this issue, others to explore are:
(1) restricting executionEnvID to valid pod names: this disallows uppercase characters, etc which IMO seems unnecessarily restrictive.
(2) using a hash of the executionEnvID as a pod prefix: this makes it VERY difficult to visually link pods with environments.

## Test Plan
Added a variety of unit tests and tested in various scenarios.

## Rollout Plan (if applicable)
May be rolled out immediately.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/COR-1625/sanitize-actor-replica-pod-names

## Checklist
* [x] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation