Update various package versions with vulnerabilities (#462)
## Overview
Update package versions for
* github.com/Azure/azure-sdk-for-go/sdk/azidentity
* github.com/go-jose/go-jose/v3
* golang.org/x/net

Note: there is a vulnerability in `gopkg.in/square/go-jose.v2` pulled in indirectly for a bunch of tests without a fix. Not tackling this one

## Test Plan
- [x] Run functional and api tests on Union

## Rollout Plan (if applicable)
Merge and pick up in unionai/cloud

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
ref COR-1764