Handle no container status yet (#466)
## Overview
Fixes a race in populating task logs and pod `Container` or `ContainerStatuses` not being available.

## Test Plan
Added unittests

## Rollout Plan (if applicable)
Merge, then bring to unionai/cloud

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS