Add extra template vars for task start and end time (#471)
## Overview
Start populating `{{.taskStartTime}}` and `{{.taskEndTime}}` template vars for fasttask logs. Drop pod time fields, as they are not relevant for task lifetimes (or accurate at this point)

## Test Plan
- [x] Deployed to dogfood-gcp and [verified link](https://dogfood-gcp.cloud-staging.union.ai/console/projects/flytesnacks/domains/development/executions/a6xnkfk24vvtwhdvvjhh/nodeId/n0/nodes)
- [x] Deployed to dogfood and [verified link](https://dogfood.cloud-staging.union.ai/console/projects/flytesnacks/domains/development/executions/ajvfw8chz7xlrr9ck787/nodeId/n0/nodes) 

## Rollout Plan (if applicable)
Bring to cloud and incorporate into fasttask log link templates

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Tracking Issue(s)
ref COR-1910