Add task exec id to log inputs (#476)
## Overview
Include `TaskExecutionID` in the log template vars so that these can be used in log templates. This matches what we do for the k8s plugin

## Test Plan
- [x] Deployed to dogfood and verified that a template with these vars works

## Rollout Plan (if applicable)
Merge and bring to unionai/cloud

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
ref COR-1433