Add Is Mutable Attribute Source of Truth Reason (#513)
## Overview
This PR focuses on refactoring the GetConfiguration endpoint to include the reason when an attribute is immutable.

Additionally, I identified an issue in how we handle the **default** configuration. Currently, we incorrectly use the domain configuration as the default. Instead, when retrieving the default configuration for a project-domain level configuration, it should include all configurations lower than the project-domain level, including project, org, and domain configurations. A fix for this issue is included in this PR.

## Test Plan
Deployed to dogfood-gcp and can see
```json
{
  "source": "DOMAIN",
  "value": {
    "attributes": {
      "projectQuotaCpu": "4096",
      "projectQuotaMemory": "2Ti",
      "projectQuotaNvidiaGpu": "256"
    }
  },
  "isMutable": {
    "value": false,
    "reason": "Due to GCP architecture, resource quotas must be changed at the cluster level."
  }
}
```
and if I remove my admin permission, i can see
```json
{
  "source": "GLOBAL",
  "value": null,
  "isMutable": {
    "value": false,
    "reason": "You don't have the corresponding permission to edit the attributes."
  }
}
```

If I set `OnlyGetLowerLevelConfiguration` to true, I can see the correct result.

## Rollout Plan (if applicable)
Needs to rollout in coordination with @nathandpeterson's change https://github.com/unionai/cloud/pull/9141

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
fixes COR-2083
fixes COR-1742