Adds fast-register directory to the top of python path in actors (#516)
## Overview
Currently, the fast execution directory is added to the bottom of the PATH. That means files in `/root` will have a higher precedence. For example, lets say `/root/task.py` exists and `task.py` is in the fast-register tarball. The old `/root/task.py` will be loaded first. This PR moves the fast-register code at the top of the python path.

## Test Plan
Ran new actor code with code already in `/root` and fast-register with files with the same name. With this change, the new fast-register code is picked up.

## Rollout Plan (if applicable)
Ship with `union`.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS