Add explainer on merged PRs for upstreamable commits (#498)
## Overview
Add a workflow to comment on merged PRs with the `[x] To be upstreamed to OSS` box checked with steps for manually upstreaming the commit to flyteorg/flyte. This is in lieu of good automation options here right now and attempts to streamline this process as much as possible.

https://github.com/unionai/flyte/pull/497
<img width="927" alt="Screenshot 2024-09-23 at 8 56 45 PM" src="https://github.com/user-attachments/assets/9db6a2b3-68b5-471b-b1fd-a4fca20905a4">

## Test Plan
Merged a dummy commit and observed the PR was commented on

## Rollout Plan (if applicable)
Merge **after** batch upstreaming all outstanding commits.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS