Check base prefix lib as well for python module files (#532)
## Overview
When an image uses a virtual environment with system python, there are two base prefixes. One for the system python and another for the virtualenv. We need to check both places to ensure that no python library code gets unloaded.

## Test Plan
Test with image using a virtual env based on the system python.

## Rollout Plan (if applicable)
Ship another version of unionai

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS