Propeller GC minute granularity (#550)
## Overview
In order to support more execution throughput we need to be able to clean completed workflow CRDs quicker. This PR introduces a 5-minute granular support for MaxTTL propeller config.

## Test Plan
* tested via unit tests and local setup

## Rollout Plan (if applicable)
* integrate changes to cloud and redeploy managed-clusters

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
* fixes COR-2256

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation