Add new 'Workspace Generated' status to named entities for workspace task definitions (#545)
## Overview
In addition to the changes in the title, the UI already adds an active state filter when querying named entities to power the 'Tasks' view in the console, but for CLI and other callers it's nice to add this this implicit filter like we do to exclude non-active workspace definitions in list calls.

## Test Plan
Test on staging, verified workflow task definitions didn't show up in the console.

## Rollout Plan (if applicable)
Merge into Union Cloud, roll out serverless control planes via org-sync

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS
No, but the enum fields are reserved in https://github.com/flyteorg/flyte/pull/5975

## Issue
ref cor-2257

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation