Update golang version to 1.22 (#554)
## Overview
Downstream https://github.com/flyteorg/flyte/pull/5032 and https://github.com/flyteorg/flyte/pull/5723 to update golang to 1.22.

This is in preparation for bringing the Flyte 1.13.3.

## Test Plan
Relying on unit tests for now. Will test in a tenant.

## Rollout Plan (if applicable)
N/A

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
Part of https://linear.app/unionai/issue/COR-1861/upgrade-to-go122

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation