Revert "[COR-2297/] Fix nested offloaded type validation (#552) (#5996)" (#6045) (#572)
## Overview
Cherry-pick https://github.com/flyteorg/flyte/pull/6045

## Test Plan
Already verified in OSS

## Rollout Plan (if applicable)
N/A

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
Related to https://linear.app/unionai/issue/COR-2297/nested-collection-of-offloaded-literal-fails-when-using-launchplans

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation