Use default podTemplate for the ray submitter pod (#586)
## Overview
Currently, the resources allocated to the Ray submitter pod match the resources specified in @task. However, the submitter pod doesn’t require many resources, as it only sends an HTTP request to the head node. We can create a default pod template for the submitter with 500m CPU, which should be sufficient.

## Test Plan
Run a rayjob and check the submitter pod resource.

## Rollout Plan (if applicable)
staging -> canary -> production

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/DX-1207/use-default-podtemplate-for-the-ray-submitter-pod

## Checklist
* [x] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation