Special-case type annotations to force cache invalidations for msgpack-binary literals (#587)
## Overview
In order to ease the transition to Flyte 1.14 and the new support for dataclasses+pydantic models and cached tasks we downstream https://github.com/flyteorg/flyte/pull/6078.

This change affects the caching of tasks that use flytekit 1.14 and contain dataclass or pydantic models as inputs or outputs. 

I need to double-check that the cache service uses a similar algorithm to generate the signature of a task [as the OSS one](https://github.com/flyteorg/flyte/blob/master/flytepropeller/pkg/controller/nodes/catalog/datacatalog/transformer.go#L85-L119).

## Test Plan
Unit tests and will test on a tenant.

## Rollout Plan (if applicable)
 No special consideration.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
*TODO: Link Linear issue(s) using [magic words](https://linear.app/docs/github#magic-words). `fixes` will move to merged status, while `ref` will only link the PR.*

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation