Use base_prefix and prefix in actor executor (#589)
## Overview
For some platforms, the "lib" could be "lib64" or `"lib"`. It's safer to just check the `sys.base_prefix` and `sys.prefix` instead. `os.path.dirname(sys.executable)` is not longer needed because it is included in the `sys.prefix`.

## Test Plan
The current tests in `fasttask/worker/functional-tests/test_fast_register_actor.py` already checks this code path.

## Rollout Plan (if applicable)
Ship a new version of unionai-actor

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS