Set ActorEnvironment TTL to 30s in functional test (#617)
## Overview
ActorEnvironment TTL was set to 10s in the functional tests. However, [this PR](https://github.com/unionai/unionai/pull/604) introduced a lower-bound of 30s, causing a testing failure.

## Test Plan
If the functional tests on this PR succeed, then we are good.

## Rollout Plan (if applicable)
This is a testing only change, so no need to rollout.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/COR-2658/fasttask-functional-tests-are-failing-with-actor-ttl

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation