Remove noisy token fetch log lines (#619)
## Overview
This change removes and or reduces level of a bunch of log lines in the fetch token path that were added as part of the expiration investigation

Note today on buildkite agents we fail to save the token to the keyring because the key ring is not installed

![Screenshot 2025-01-13 at 11 04 29 AM](https://github.com/user-attachments/assets/1b3b9ff0-b8c4-44c8-9cdf-ae70d7de91c2)


## Test Plan
- [x] Run api tests and observe reduced spam ([run](https://buildkite.com/unionai/api-tests-staging-run/builds/322#01946151-1de2-4005-b3a1-f345ec296ec3) -- failures expected)

## Rollout Plan (if applicable)
Merge and bring to unionai/cloud

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation