Add a property bag to node metadata (#613)
## Overview
This allows plugins to pass execution instructions to the backend plugin without having to introduce new protos..

## Test Plan
Deploy to a staging cluster and test passing cluster_pool to enforce running an LP on some other clusterpool

## Rollout Plan (if applicable)
Deploy internal tenant and test

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/COR-2676/allow-passing-cluster-pool-at-node-and-lp-levels

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation