Use ClusterPoolAssignment from LaunchPlan or Node.Config when executing LPs (#615)
## Overview
This change uses ClusterPoolAssignment from the LP or from Node.Config if set when propeller executes a launchplan

## Test Plan
Deploy in dogfood and attempt to run one LP in dogfood-1 and have it launch another LP in dogfood-2

## Rollout Plan (if applicable)
Merge and update cloud repo

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
*TODO: Link Linear issue(s) using [magic words](https://linear.app/docs/github#magic-words). `fixes` will move to merged status, while `ref` will only link the PR.*

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation