add INITIALIZING state to actor environment management (#623)
## Overview
Currently, if all of the `createPods` calls fail when creating an environment, the environment is still maintained in the in-memory store. This allows other tasks to request and retrieve it (updating the TTL expiration). However, the other tasks will be stuck in a "no workers available" state indefinitely (because creating the replica Pods fails). The result, is that when two (or more) tasks execute in relative unison, if the replica Pod creation will never succeed (ex. incorrect volume mount) there is the _potential_ that the system gets stuck continually attempting to create replica Pods rather than fast failing.

This PR introduces an INITIALIZING state to ensure cleanup of environments if creating initial pods fails.

## Test Plan
This was tested locally.

## Rollout Plan (if applicable)
May be rolled out immediately.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/COR-2651/auto-fail-before-adding-actor-environment-if-failed-to-create-pods

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation