fixing lint on stripping metadata from dataclass types (#626) (#6177) (#631)
## Overview
Fixing the[ lint errors](https://github.com/flyteorg/flyte/pull/6177) from upstreaming [this PR](https://github.com/unionai/flyte/pull/626). This means that OSS and the fork remain consistent.

## Test Plan
This was tested as part of the original work.

## Rollout Plan (if applicable)
This PR does not effect functionality at all, it is purely cosmetic.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/COR-2720/strip-metadata-from-dataclass-types-in-flyteworkflow-crd-interfaces

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation