Unload all new modules between actor task executions (#649)
## Overview
This PR updates the Actor implementation to delete all new modules from `sys.modules` between task executions. This ensures modules are not re-used between task executions, potentially causing older versions of code to be executed.

## Test Plan
This was tested locally by printing modules before and after execution to ensure correct deletion.

## Rollout Plan (if applicable)
As with all Actor updates, this will need to be pulled into the `unionai/unionai` repo and a new Actor release. The plan is to release this as an "alpha" and test with MyHeritage to ensure correctness before completely rolling out.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/COR-2811/actor-second-run-fails-on-python-import

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation