updating actors to default log-level to warn (#652)
## Overview
This PR updates the default log-level for actors to `warn` rather than `info`. This `info` level was breaking API functional tests on Union deploys. We will take up re-enabling this in figure work.

## Test Plan
This was tested locally to ensure correct logging output.

## Rollout Plan (if applicable)
The functional tests on deploy will be executed to ensure logging API tests pass.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/COR-2830/actor-default-to-warn-debug-level

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation