Override literal in the case of attribute access of primitive values  (#668)
## Overview
As the title suggests, we need to provide better support for attribute access in the case of primitives. This downstreams https://github.com/flyteorg/flyte/pull/6194.

## Test Plan
Tested extensively in OSS.

## Rollout Plan (if applicable)
Revert PR

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
fixes https://linear.app/unionai/issue/DX-1135/implement-backup-solution-of-turning-primitive-types-into-literals

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation